Split device.cpp into device.cpp and device_manager.cpp
The original device.cpp mixed the pure class implementation (constructor,
register_sysfs_path, measurement virtuals, collect_json_fields) with
higher-level manager logic (create_all_devices, clear_all_devices,
devices_start/end_measurement, report_devices, show_report_devices) that
pulls in all device subtypes plus the report/, display/, and measurement/
subsystems.

Split into:
  device.cpp         - class device methods + all_devices vector only;
                       minimal includes (device.h, limits.h, stdlib.h,
                       unistd.h) — safe to link in unit tests
  device_manager.cpp - everything else; keeps the heavy include chain

Unit tests now link device.cpp directly instead of a hand-written stub,
reducing the stub to just three missing symbols from measurement/ and
devlist/ (global_power, save_all_results, register_devpower).
